Atualizações nos projétis do jogo, agora eles dão dano nos inimigos e tem um tempo de recarga entre um tiro e outro

+++ src/helpers/Constants.java
@@ -39,7 +39,7 @@ public class Constants {
             return "";
         }
 
-        public static float GetStartDmg(int towerType) {
+        public static int GetStartDmg(int towerType) {
             switch (towerType) {
                 case CANNON:
                     return 25;
